+++ src/constants/patch.ts
@@ -1,5 +1,2 @@
 export const commitQueueAlias = "__commit_queue";
 export const commitQueueRequester = "merge_test";
-
-export const mergeTaskName = "merge-patch";
-export const mergeTaskVariant = "commit-queue-merge";
